requirements.txt
pybit==5.10.1
python-telegram-bot==22.0
python-dotenv==1.1.0
python-telegram-bot[job-queue]==22.0